--- src/API/actionTypes.ts
@@ -1,4 +1,0 @@
-export const LOADING_START = "LOADING_START";
-export const LOADING_END = "LOADING_END";
-export const LOGIN = "LOGIN";
-export const LOGOUT = "LOGOUT";
