+++ src/Redux/types.ts
@@ -2,3 +2,4 @@ export const LOADING_START = "LOADING_START";
 export const LOADING_END = "LOADING_END";
 export const LOGIN = "LOGIN";
 export const LOGOUT = "LOGOUT";
+export const TEST_ACTION = "TEST_ACTION";
